Making extended example (2) work
Several changes towards a a working 2nd data example.

* the version 2 file format includes the species alpha codes (e.g.
"lade") instead of fixed numbers.
In order to enable this, species codes/ids need to be available during
file load. This required some plumbing.

* project areas are not always rectangles: example 2 is a rectangle, and
this brought up some issues

* performance is pretty bad for larger projects: I updated some of the
code that generates statistics (I added also some timing code that can
be removed again)

* it still crashes in release mode - I tried some things to track down
the error (e.g., exception handling), but have not found the root cause

* in terrain.cpp I changed *experimentally* the reading of elevation
files to the same sequence as is commonly used by GIS software (i.e., in
rows, starting from the top). The elv file in the example is based on the
GIS grid (ESRI ascii format) and *not* the custom format of the
elevation file

* added some files as a second example (data/test2). These are
just five years of the baseline run and the custom-elevation file (
which is too big for GitHub, therefore zipped).