Mouse-based selection of overview region
Implemented mouse-based region selection (with right mouse button) and translation (with left mouse button). I have also centered the newly selected region in the 3D view. Some problems remain:
(1) the 3D view needs to be zoomed to encompass the new terrain properly
(2) On Apple the overview window goes black after a new region is created. This seems to have to do with the destruction and recreation of the perspectiveView.
(3) There are grid artefacts on the landscape. Although, these might always have been there, but are now visible because we see the landscape in more detail.