Floating menu
Moved the overview map on top of the perspective view. This required a fair amount of rewiring. The behaviour where the overview map disappears when the main window is being moved is by design. The alternative led to lag with the positioning of the overview map. There are still some niggly issues remaining: resizing the window messes up the alignment and right clicking on the bar of the main window causes the overview map to disappear even when the mouse is released.