Fix for ecoystem offset
Note that this fix has required an update to the pdb and elv file format. PDB files are now on version 3.0. The change simply adds a coordinate for the top left corner of the terrain. This way the eocsystem and terrain can have diferent global coordinates and the ecoystem can be shifted correctly on the terrain. This requires new data, which can be downloaded from: https://drive.google.com/file/d/14ERGaRgNoO9JqLJNgYbXcnFCnpPbOH3X/view?usp=share_link