Auto-scaling of DBH distribution and performance
DBH distributions now scale correctly to the maximum of a time step.
Improved the startup performance significantly by kicking saplings out
of calculating statistics. Saplings require full new sampling of trees /
positions which is expensive. For the graphs it is imho ok to omit
saplings. There are other issues with the graphs, e.g. that the *total
area* of the rectangle is used for calculating per ha values (instead of
forested area).